+++ previous-versions/sync-for-commerce-version-1/pkg/models/operations/createaccountingpayment.go
@@ -50,8 +50,6 @@ type CreateAccountingPaymentResponse struct {
 	AccountingCreatePaymentResponse *shared.AccountingCreatePaymentResponse
 	// HTTP response content type for this operation
 	ContentType string
-	// The request made is not valid.
-	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
@@ -70,13 +68,6 @@ func (o *CreateAccountingPaymentResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
-}
-
-func (o *CreateAccountingPaymentResponse) GetErrorMessage() *shared.ErrorMessage {
-	if o == nil {
-		return nil
-	}
-	return o.ErrorMessage
 }
 
 func (o *CreateAccountingPaymentResponse) GetStatusCode() int {