+++ previous-versions/sync-for-commerce-version-1/pkg/models/shared/accountingcreditnote.go
@@ -111,7 +111,7 @@ type AccountingCreditNote struct {
 	// Any additional information about the credit note. Where possible, Codat links to a data field in the accounting platform that is publicly available. This means that the contents of the note field are included when a credit note is emailed from the accounting platform to the customer.
 	Note *string `json:"note,omitempty"`
 	// An array of payment allocations.
-	PaymentAllocations []PaymentAllocationsitems `json:"paymentAllocations,omitempty"`
+	PaymentAllocations []PaymentAllocationItems `json:"paymentAllocations,omitempty"`
 	// Unused balance of totalAmount originally raised.
 	RemainingCredit    *decimal.Big `decimal:"number" json:"remainingCredit"`
 	SourceModifiedDate *string      `json:"sourceModifiedDate,omitempty"`
@@ -129,7 +129,7 @@ type AccountingCreditNote struct {
 	TotalDiscount *decimal.Big `decimal:"number" json:"totalDiscount"`
 	// Any tax applied to the credit note amount.
 	TotalTaxAmount *decimal.Big          `decimal:"number" json:"totalTaxAmount"`
-	WithholdingTax []WithholdingTaxitems `json:"withholdingTax,omitempty"`
+	WithholdingTax []WithholdingTaxItems `json:"withholdingTax,omitempty"`
 }
 
 func (a AccountingCreditNote) MarshalJSON() ([]byte, error) {
@@ -241,7 +241,7 @@ func (o *AccountingCreditNote) GetNote() *string {
 	return o.Note
 }
 
-func (o *AccountingCreditNote) GetPaymentAllocations() []PaymentAllocationsitems {
+func (o *AccountingCreditNote) GetPaymentAllocations() []PaymentAllocationItems {
 	if o == nil {
 		return nil
 	}
@@ -304,7 +304,7 @@ func (o *AccountingCreditNote) GetTotalTaxAmount() *decimal.Big {
 	return o.TotalTaxAmount
 }
 
-func (o *AccountingCreditNote) GetWithholdingTax() []WithholdingTaxitems {
+func (o *AccountingCreditNote) GetWithholdingTax() []WithholdingTaxItems {
 	if o == nil {
 		return nil
 	}